+++ src/fondant/dataset/__init__.py
@@ -3,7 +3,7 @@ from .lightweight_component import (  # noqa
     LightweightComponent,
     lightweight_component,
 )
-from .pipeline import (  # noqa
+from .dataset import (  # noqa
     VALID_ACCELERATOR_TYPES,
     VALID_VERTEX_ACCELERATOR_TYPES,
     ComponentOp,
