+++ tests/examples/example_modules/invalid_double_pipeline.py
@@ -1,4 +1,4 @@
-from fondant.pipeline import Pipeline
+from fondant.dataset import Pipeline
 
 TEST_PIPELINE = Pipeline(name="test_pipeline", base_path="some/path")
 TEST_PIPELINE_2 = Pipeline(name="test_pipeline", base_path="some/path")
