+++ tests/pipeline/test_compiler.py
@@ -18,14 +18,14 @@ from fondant.core.component_spec import ComponentSpec
 from fondant.core.exceptions import InvalidPipelineDefinition
 from fondant.core.manifest import Manifest, Metadata
 from fondant.core.schema import CloudCredentialsMount
-from fondant.pipeline import (
+from fondant.dataset import (
     ComponentOp,
     Dataset,
     Pipeline,
     Resources,
     lightweight_component,
 )
-from fondant.pipeline.compiler import (
+from fondant.dataset.compiler import (
     DockerCompiler,
     KubeFlowCompiler,
     KubeflowComponentSpec,
